refactor: Split up reports.py
