+++ src/ctk_functions/data/pyrite_template/pyrite_tscore_distribution.png
