+++ src/ctk_functions/routers/pyrite/controller.py
@@ -11,10 +11,10 @@ from fastapi import status
 
 from ctk_functions.core import config
 from ctk_functions.microservices.sql import models
+from ctk_functions.routers.pyrite import sql_data
 from ctk_functions.routers.pyrite.reports import reports
 from ctk_functions.routers.pyrite.tables import (
     base,
-    utils,
 )
 
 logger = config.get_logger()
@@ -83,7 +83,9 @@ class PyriteReport:
         sanitized_mrn = self._mrn.replace("\r", "").replace("\n", "")
         logger.debug("Fetching participant %s.", sanitized_mrn)
         try:
-            return utils.fetch_participant_row("MRN", self._mrn, models.CmiHbnIdTrack)  # type: ignore[no-any-return, unused-ignore] # Getting errors both when no-any-return is, and is not used.
+            return sql_data.fetch_participant_row(
+                "MRN", self._mrn, models.CmiHbnIdTrack
+            )  # type: ignore[no-any-return, unused-ignore] # Getting errors both when no-any-return is, and is not used.
         except base.TableDataNotFoundError as exception_info:
             raise fastapi.HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
