+++ src/ctk_functions/routers/pyrite/reports/introduction.py
@@ -1,11 +1,17 @@
 """Contains dataclasses to build the introduction page."""
 
+from collections.abc import Iterable
 from typing import cast
 
 import pydantic
 
+from ctk_functions.core import config
 from ctk_functions.microservices.sql import models
-from ctk_functions.routers.pyrite.reports import utils
+from ctk_functions.routers.pyrite import sql_data, types
+from ctk_functions.routers.pyrite.reports import sections
+
+settings = config.get_settings()
+DATA_DIR = settings.DATA_DIR
 
 
 @pydantic.dataclasses.dataclass(frozen=True)
@@ -20,7 +26,7 @@ class TableColumn:
 class TestOverview:
     """Definition of the introduction overview of a test."""
 
-    id: utils.TestId
+    id: types.TestId
     title: str
     description: str
     date_table: TableColumn | None
@@ -30,7 +36,7 @@ class TestOverview:
 class TestOverviewManager:
     """Dataclass for all test descriptions."""
 
-    def fetch(self, test_id: utils.TestId) -> TestOverview | None:
+    def fetch(self, test_id: types.TestId) -> TestOverview | None:
         """Convenience method that mimics getattr with correct typing."""
         try:
             return cast("TestOverview", getattr(self, test_id))
@@ -67,9 +73,11 @@ class TestOverviewManager:
         description="",
         date_table=TableColumn(model=models.SummaryScores, column="CTOPP_Date"),
     )
+
     grooved_pegboard = TestOverview(
         id="grooved_pegboard",
         title="Lafayette Grooved Pegboard Test",
+        description="",
         date_table=None,
     )
 
@@ -150,4 +158,103 @@ class TestOverviewManager:
             "Figure Weights, Digit Span, Picture Memory, Coding, Symbol Search"
         ),
         date_table=TableColumn(model=models.SummaryScores, column="WISC_Date"),
+    )
+
+
+def test_ids_to_introduction(
+    mrn: str, test_ids: Iterable[types.TestId]
+) -> tuple[sections.Section, ...]:
+    """Creates the introduction section of the report.
+
+    Args:
+        mrn: The participant's unique identifier.
+        test_ids: The IDs of the tests to include.
+
+    Returns:
+        The introduction section of the report.
+    """
+    overviews = TestOverviewManager()
+    used_overviews = [overviews.fetch(test_id) for test_id in test_ids]
+    introduction_sections = [
+        _overview_to_sections(mrn, overview) for overview in used_overviews if overview
+    ]
+    return (
+        sections.ParagraphSection(
+            content="STANDARDIZED TESTING, INTERVIEW AND QUESTIONNAIRE RESULTS",
+            style="Heading 1",
+        ),
+        sections.ParagraphSection(
+            content=(
+                "This report provides a summary of the following tests, "
+                "questionnaires, and clinical interviews that were administered during "
+                "participation in the study. Areas assessed include general cognitive "
+                "ability, academic achievement, language and fine motor coordination. "
+                "Clinical interviews and questionnaires were used to assess social, "
+                "emotional and behavioral functioning. Please reference Appendix A for "
+                "a full description of the measures administered in the Healthy Brain "
+                "Network research protocol."
+            ),
+            style=None,
+        ),
+        *introduction_sections,
+        sections.PageBreak(),
+        sections.ParagraphSection(
+            content="What do the scores represent?",
+            style="Heading 2",
+        ),
+        sections.ParagraphSection(
+            content=(
+                "Standard scores, T scores, and percentile ranks can indicate "
+                "{{FIRST_NAME_POSSESSIVE}} performance compared to other children in "
+                "the same age or same grade (see Figure below). A standard score of "
+                "100 is the mean of the normative sample (individuals in the same age "
+                "or same grade), and a standard score within the range of 90-109 "
+                "indicates that an individual's performance or rating is within the "
+                "average or typical range. A T score of 50 is the mean of the "
+                "normative sample, and a T score within the range of 43-57 indicates "
+                "that an individual's performance or ratings is within the average "
+                "range. A percentile rank of 50 is the median of the normative sample, "
+                "meaning that an individual's performance equals or exceeds 50% of the "
+                "same-age peers. A percentile rank within the range of 25-75 indicates "
+                "that an individual's performance or rating is within the average."
+            ),
+            style=None,
+        ),
+        sections.ImageSection(path=DATA_DIR / "pyrite_tscore_distribution.png"),
+    )
+
+
+def _overview_to_sections(mrn: str, overview: TestOverview) -> sections.Section:
+    """Converts a test description to a section.
+
+    Args:
+        mrn: The participant's unique identifier.
+        overview: The description of the test.
+
+    Returns:
+        The section for this test..
+    """
+    if overview.date_table:
+        data = sql_data.fetch_participant_row(
+            "person_id", mrn, overview.date_table.model
+        )
+        administer_date = str(getattr(data, overview.date_table.column))
+    else:
+        administer_date = ""
+
+    if not overview.description:
+        texts: tuple[str, ...] = (overview.title + "\n", "Administered on: ")
+        run_styles: tuple[None | sections.RunStyles, ...] = (None, None)
+    else:
+        texts = (
+            overview.title + "\n",
+            overview.description + "\n",
+            f"Administered on: {administer_date}",
+        )
+        run_styles = (None, sections.RunStyles.Emphasis, None)
+
+    return sections.RunsSection(
+        content=texts,
+        paragraph_style="Normal Hanging",
+        run_styles=run_styles,
     )
