+++ src/ctk_functions/routers/pyrite/tables/base.py
@@ -14,9 +14,9 @@ from docx import document, shared, table
 from docx.enum import text
 from docx.text import paragraph
 
-import ctk_functions.routers.pyrite.reports.utils
 from ctk_functions.core import config
 from ctk_functions.microservices.sql import models
+from ctk_functions.routers.pyrite import types
 
 logger = config.get_logger()
 
@@ -441,9 +441,7 @@ class DataProducer(abc.ABC):
 
     @classmethod
     @abc.abstractmethod
-    def test_ids(
-        cls, mrn: str
-    ) -> tuple[ctk_functions.routers.pyrite.reports.utils.TestId, ...]:
+    def test_ids(cls, mrn: str) -> tuple[types.TestId, ...]:
         """The IDs of the tests used to produce this data.
 
         Args:
