+++ src/ctk_functions/routers/pyrite/tables/ctopp2.py
@@ -3,9 +3,9 @@
 import dataclasses
 import functools
 
-import ctk_functions.routers.pyrite.reports.utils
 from ctk_functions.microservices.sql import models
-from ctk_functions.routers.pyrite.tables import base, utils
+from ctk_functions.routers.pyrite import sql_data, types
+from ctk_functions.routers.pyrite.tables import base
 
 
 @dataclasses.dataclass
@@ -34,9 +34,7 @@ class _Ctopp2DataSource(base.DataProducer):
     """Fetches the data for the CTOPP-2 table."""
 
     @classmethod
-    def test_ids(
-        cls, mrn: str
-    ) -> tuple[ctk_functions.routers.pyrite.reports.utils.TestId, ...]:  # noqa: ARG003
+    def test_ids(cls, mrn: str) -> tuple[types.TestId, ...]:  # noqa: ARG003
         return ("ctopp_2",)
 
     @classmethod
@@ -50,7 +48,7 @@ class _Ctopp2DataSource(base.DataProducer):
         Returns:
             The text contents of the Word table.
         """
-        data = utils.fetch_participant_row("person_id", mrn, models.SummaryScores)
+        data = sql_data.fetch_participant_row("person_id", mrn, models.SummaryScores)
         header = ("CTOPP - 2 Rapid Naming", "Number of Errors")
         content = [
             (label.name, getattr(data, label.score_column))
