+++ src/ctk_functions/routers/pyrite/tables/grooved_pegboard.py
@@ -3,8 +3,8 @@
 import dataclasses
 import functools
 
-import ctk_functions.routers.pyrite.reports.utils
 from ctk_functions.microservices.sql import models
+from ctk_functions.routers.pyrite import sql_data, types
 from ctk_functions.routers.pyrite.tables import base, utils
 
 
@@ -32,9 +32,7 @@ class _GroovedPegboardDataSource(base.DataProducer):
     """Fetches the data for the Grooved Pegboard table."""
 
     @classmethod
-    def test_ids(
-        cls, mrn: str
-    ) -> tuple[ctk_functions.routers.pyrite.reports.utils.TestId, ...]:  # noqa: ARG003
+    def test_ids(cls, mrn: str) -> tuple[types.TestId, ...]:  # noqa: ARG003
         return ("grooved_pegboard",)
 
     @classmethod
@@ -48,7 +46,7 @@ class _GroovedPegboardDataSource(base.DataProducer):
         Returns:
             The text contents of the Word table.
         """
-        data = utils.fetch_participant_row("EID", mrn, models.GroovedPegboard)
+        data = sql_data.fetch_participant_row("EID", mrn, models.GroovedPegboard)
         header = ("Hand", "Z-Score", "Percentile", "Range")
         content_rows = [
             _create_pegboard_content_row(label, data) for label in PEGBOARD_ROW_LABELS
