+++ src/ctk_functions/routers/pyrite/tables/language.py
@@ -6,8 +6,8 @@ from typing import Literal
 
 from docx import shared
 
-import ctk_functions.routers.pyrite.reports.utils
 from ctk_functions.microservices.sql import models
+from ctk_functions.routers.pyrite import sql_data, types
 from ctk_functions.routers.pyrite.tables import base, utils
 
 COLUMN_WIDTHS = (
@@ -108,11 +108,9 @@ class _LanguageDataSource(base.DataProducer):
     """Fetches the data for the Language table."""
 
     @classmethod
-    def test_ids(
-        cls, mrn: str
-    ) -> tuple[ctk_functions.routers.pyrite.reports.utils.TestId, ...]:
+    def test_ids(cls, mrn: str) -> tuple[types.TestId, ...]:
         tests = [row[0] for row in cls.fetch(mrn)]
-        test_ids: list[ctk_functions.routers.pyrite.reports.utils.TestId] = []
+        test_ids: list[types.TestId] = []
         if any("ctopp" in test.lower() for test in tests):
             test_ids.append("ctopp_2")
         if any("wiat" in test.lower() for test in tests):
@@ -130,7 +128,7 @@ class _LanguageDataSource(base.DataProducer):
         Returns:
             The markup for the Word table.
         """
-        data = utils.fetch_participant_row("person_id", mrn, models.SummaryScores)
+        data = sql_data.fetch_participant_row("person_id", mrn, models.SummaryScores)
         header = ("Test", "Subtest", "Standard Score", "Percentile", "Range")
 
         content_rows = [
