+++ src/ctk_functions/routers/pyrite/tables/scared.py
@@ -4,8 +4,8 @@ import functools
 
 import cmi_docx
 
-import ctk_functions.routers.pyrite.reports.utils
 from ctk_functions.microservices.sql import models
+from ctk_functions.routers.pyrite import types
 from ctk_functions.routers.pyrite.tables import base
 from ctk_functions.routers.pyrite.tables.generic import parent_child
 
@@ -107,9 +107,7 @@ class _ScaredDataSource(base.DataProducer):
     """Fetches the data for the Scared table."""
 
     @classmethod
-    def test_ids(
-        cls, mrn: str
-    ) -> tuple[ctk_functions.routers.pyrite.reports.utils.TestId, ...]:  # noqa: ARG003
+    def test_ids(cls, mrn: str) -> tuple[types.TestId, ...]:  # noqa: ARG003
         return ("scared",)
 
     @classmethod
