+++ src/ctk_functions/routers/pyrite/tables/swan.py
@@ -5,9 +5,9 @@ import functools
 
 import cmi_docx
 
-import ctk_functions.routers.pyrite.reports.utils
 from ctk_functions.microservices.sql import models
-from ctk_functions.routers.pyrite.tables import base, utils
+from ctk_functions.routers.pyrite import sql_data, types
+from ctk_functions.routers.pyrite.tables import base
 from ctk_functions.routers.pyrite.tables.scq import COLUMN_WIDTHS
 
 
@@ -63,9 +63,7 @@ class _SwanDataSource(base.DataProducer):
     """Fetches and creates the SWAN table."""
 
     @classmethod
-    def test_ids(
-        cls, mrn: str
-    ) -> tuple[ctk_functions.routers.pyrite.reports.utils.TestId, ...]:  # noqa: ARG003
+    def test_ids(cls, mrn: str) -> tuple[types.TestId, ...]:  # noqa: ARG003
         return ("swan",)
 
     @classmethod
@@ -79,7 +77,7 @@ class _SwanDataSource(base.DataProducer):
         Returns:
             The text contents of the Word table.
         """
-        data = utils.fetch_participant_row("EID", mrn, models.Swan)
+        data = sql_data.fetch_participant_row("EID", mrn, models.Swan)
         header = ("Subscale", "Score", "Clinical Relevance")
 
         # Scores are clipped at 0, the total needs to be adjusted for clipped scores.
